Copilot adoption: split the models file and pin the reporting window
Two changes to Common.Entities.CopilotAdoption, both behaviour-preserving except the window fix.

Reorganisation (pure moves, no code edits). CopilotAdoptionModels.cs held 29 unrelated types in 1,486 lines. Split along its existing #region seams into one file per concern (options, licence types, bands, per-user rows, agents, unlicensed, concentration, summary), and lifted CopilotAdoptionAnalysis out of the 1,320-line service file so the feature's main result type is findable. Namespace is unchanged, so no call site moves. A verification pass confirmed all 29 type declarations and every code line carried over before anything was written. The remaining large files are each a single cohesive class and were deliberately left alone - splitting those into partials would scatter one class across files rather than help.

Window fix. 'Last N days' started at midnight N days ago, which spans N+1 distinct calendar dates. A user active every single day therefore recorded 29 active days in a '28-day' window while the frequency component divides by a target derived from 28 - inflating the score of precisely the most engaged users, and making the dates shown on screen a day wider than their own label. The window definition now lives in CopilotAdoptionScoring.WindowStartUtc, next to TargetActiveDays, so the numerator's window and the denominator's target cannot drift apart. The agent inventory's reported HistoryDays had the same off-by-one and is fixed with it.

Tests: 2 new regression tests pinning the window to exactly N dates. Copilot adoption suite 109/109.


Copilot-Session: 87347b97-407c-4dbe-bae5-e8880d0fc77d
