+++ src/AnalyticsEngine/Common/Entities/Entities.csproj
@@ -83,13 +83,21 @@
     <Compile Include="BuildConstants.cs" />
     <Compile Include="Config\AppConfig.cs" />
     <Compile Include="Config\AppConnectionStrings.cs" />
+    <Compile Include="CopilotAdoption\AdoptionBand.cs" />
+    <Compile Include="CopilotAdoption\CopilotAdoptionAgentModels.cs" />
+    <Compile Include="CopilotAdoption\CopilotAdoptionAnalysis.cs" />
+    <Compile Include="CopilotAdoption\CopilotAdoptionConcentrationModels.cs" />
     <Compile Include="CopilotAdoption\CopilotAdoptionExports.cs" />
-    <Compile Include="CopilotAdoption\CopilotAdoptionModels.cs" />
+    <Compile Include="CopilotAdoption\CopilotAdoptionOptions.cs" />
     <Compile Include="CopilotAdoption\CopilotAdoptionScoring.cs" />
     <Compile Include="CopilotAdoption\CopilotAdoptionService.cs" />
     <Compile Include="CopilotAdoption\CopilotAdoptionSql.cs" />
+    <Compile Include="CopilotAdoption\CopilotAdoptionSummaryModels.cs" />
+    <Compile Include="CopilotAdoption\CopilotAdoptionUnlicensedModels.cs" />
+    <Compile Include="CopilotAdoption\CopilotAdoptionUserModels.cs" />
     <Compile Include="CopilotAdoption\CopilotAdoptionWorkbook.cs" />
     <Compile Include="CopilotAdoption\CopilotLicenceClassifier.cs" />
+    <Compile Include="CopilotAdoption\CopilotLicenceModels.cs" />
     <Compile Include="CopilotAdoption\CsvSerialiser.cs" />
     <Compile Include="Config\ImportConfig.cs" />
     <Compile Include="Config\UserGroupsFilterModel.cs" />
